app/requirements.txt
SpeechRecognition==3.10.4
pyttsx3==2.90
mysql-connector-python==8.0.21
TextBlob==0.15.3
PyAudio==0.2.11
sounddevice==0.4.5
numpy==1.21.0
scipy==1.7.0
Flask>=2.1.0
Werkzeug==2.0.3
pandas




